automatically backed up by learn

+++ app.rb
@@ -21,8 +21,8 @@ class App < Sinatra::Base
   # Code your final two routes here:
 
   get '/multiply/:one/:two' do
-    number1 = params[:one].to_i
-    number2 = params[:two].to_i
+    number1 = params[:one].to_f
+    number2 = params[:two].to_f
     number1 * number2
   end
